yolo-grpc/requirements.txt
grpcio==1.58.0
grpcio-tools==1.58.0
numpy==1.26.0
opencv-python-headless==4.8.0.76
protobuf==4.24.3
